Fix JS regex for looking up csrfp_token value
We encountered a customer issue where their iOS devices were hitting a CSRF
protector error.

After some debugging, we discovered they were using a feature of the
Barracuda Web Application Firewall that signs cookies to prevent user
tampering. The signed value is saved in another cookie named
`BNES_csrfp_token`. Due to the way the regular expression is written,
it's possible for the wrong cookie to be chosen.