Use getallheaders() instead of apache_request_headers()
On Dockerized environment where apache and php are running in separate
containers, there are no `apache_` methods.

The `getallheaders()` method is an alias to the `apache_request_headers()`
method, therefore the behaviour will be identical.  The reason to use the
alias function, eh `getallheaders()` is that there are polyfills to
`getallheaders` function (like  ralouphie/getallheaders used by Guzzle,
very popular HTTP library), but there are no polyfills to the
`apache_request_headers` function.